+++ backend/src/main/java/br/univille/cofrinho/domains/frequencia/FrequenciaService.java
@@ -1,4 +1,4 @@
-package br.univille.cofrinho.domains.priority;
+package br.univille.cofrinho.domains.frequencia;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.Sort;
@@ -7,15 +7,15 @@ import org.springframework.stereotype.Service;
 import java.util.List;
 
 @Service
-public class PrioridadeService {
+public class FrequenciaService {
 
 	@Autowired
-	private PrioridadeRepository prioridadeRepository;
+	private FrequenciaRepository frequenciaRepository;
 
-	public List<PrioridadeEntity> obterTodasAsPrioridades() {
+	public List<FrequenciaEntity> obterTodasAsFrequencias(){
 		Sort ordenarCrescentePorGrandeza = Sort.by(Sort.Direction.ASC, "grandeza");
 
-		return prioridadeRepository.findAll(ordenarCrescentePorGrandeza);
+		return this.frequenciaRepository.findAll(ordenarCrescentePorGrandeza);
 	}
 
 }
