+++ backend/src/main/java/br/univille/cofrinho/domains/prioridade/PrioridadeDTO.java
@@ -1,4 +1,4 @@
-package br.univille.cofrinho.domains.priority;
+package br.univille.cofrinho.domains.prioridade;
 
 import java.util.UUID;
