+++ backend/src/main/java/br/univille/cofrinho/domains/prioridade/PrioridadeEntity.java
@@ -1,4 +1,4 @@
-package br.univille.cofrinho.domains.frequency;
+package br.univille.cofrinho.domains.prioridade;
 
 import jakarta.persistence.*;
 import org.hibernate.annotations.CreationTimestamp;
@@ -8,26 +8,26 @@ import java.util.Date;
 import java.util.UUID;
 
 @Entity
-@Table(name = "frequencia", schema = "main")
-public class FrequenciaEntity {
+@Table(name = "prioridade", schema = "main")
+public class PrioridadeEntity {
 
 	@Id
 	@GeneratedValue(strategy = GenerationType.UUID)
-	@Column(name="id")
+	@Column(name = "id")
 	private UUID id;
 
-	@Column(name="nome")
+	@Column(name = "nome")
 	private String nome;
 
 	@Column(name = "grandeza")
 	private int grandeza;
 
-	@Column(name="criado_em")
 	@CreationTimestamp
+	@Column(name = "criado_em")
 	private Date criadoEm;
 
-	@Column(name="atualizado_em")
 	@UpdateTimestamp
+	@Column(name = "atualizado_em")
 	private Date atualizadoEm;
 
 	public UUID getId() {
