+++ backend/src/main/java/br/univille/cofrinho/domains/prioridade/PrioridadeMapper.java
@@ -1,4 +1,4 @@
-package br.univille.cofrinho.domains.priority;
+package br.univille.cofrinho.domains.prioridade;
 
 public class PrioridadeMapper {
